change parsing of userInput from Integer to Double

+++ src/Account.java
@@ -3,7 +3,6 @@ import java.io.FileWriter;
 import java.io.IOException;
 import java.io.Writer;
 import java.math.BigDecimal;
-import java.math.BigInteger;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.List;
@@ -135,22 +134,22 @@ public class Account {
             }
 
                 List<String> lines = Files.readAllLines(Path.of(System.getProperty("user.dir") + "/" + Extranet.getCurrentUserName() + "/" + "acc"+ userInput + ".txt"));
-                BigDecimal moneyInAccount = BigDecimal.valueOf(Integer.parseInt(lines.get(0)));
+                BigDecimal moneyInAccount = BigDecimal.valueOf(Double.parseDouble(lines.get(0)));
                 String currencyInAccount = lines.get(1);
                 System.out.println("You have " + Extranet.getMoneyATM() + " " + Extranet.getCurrencyInATM() + " money in ATM. The chosen account have " + moneyInAccount + " " + currencyInAccount + ". Please, enter how much money do you want");
                 if (Extranet.getCurrencyInATM().equals(currencyInAccount)) {
 
                     String moneyToImport = scan.nextLine();
 
-                    if (Integer.parseInt(moneyToImport) > Extranet.getMoneyATM().intValue()) {
+                    if (Double.parseDouble(moneyToImport) > Extranet.getMoneyATM().doubleValue()) {
                         System.out.println("You want to import more money that you have in ATM");
                     } else {
                         try {
-                            BigDecimal newMoney = moneyInAccount.add(BigDecimal.valueOf(Integer.parseInt(moneyToImport)));
+                            BigDecimal newMoney = moneyInAccount.add(BigDecimal.valueOf(Double.parseDouble(moneyToImport)));
                             Writer output = new FileWriter(System.getProperty("user.dir") + "/" + Extranet.getCurrentUserName() + "/" + "acc" + userInput + ".txt", false);
                             output.write(String.valueOf(newMoney) + "\r\n" + currencyInAccount);
                             output.close();
-                            Extranet.setMoneyATM(Extranet.getMoneyATM().subtract(BigDecimal.valueOf(Integer.parseInt(moneyToImport))));
+                            Extranet.setMoneyATM(Extranet.getMoneyATM().subtract(BigDecimal.valueOf(Double.parseDouble(moneyToImport))));
                             System.out.println("Money successfully imported!");
                         } catch (Exception e) {
                             System.out.println("Something went wrong - " + e.toString());
@@ -184,7 +183,7 @@ public class Account {
             }
 
             List<String> lines = Files.readAllLines(Path.of(System.getProperty("user.dir") + "/" + Extranet.getCurrentUserName() + "/" + "acc"+ userInput + ".txt"));
-            BigDecimal moneyInAccount = BigDecimal.valueOf(Integer.parseInt(lines.get(0)));
+            BigDecimal moneyInAccount = BigDecimal.valueOf(Double.parseDouble(lines.get(0)));
             String currencyInAccount = lines.get(1);
             while(true) {
                 System.out.println("In this account you have " + moneyInAccount + " " + currencyInAccount);
@@ -193,19 +192,19 @@ public class Account {
 
                     System.out.println("how much do you want to withdraw?");
                     String moneyToWithdraw = scan.nextLine();
-                    if (Integer.parseInt(moneyToWithdraw) > Integer.parseInt(lines.get(0))) {
+                    if (Double.parseDouble(moneyToWithdraw) > Double.parseDouble(lines.get(0))) {
                         System.out.println("you have entered an amount that is higher than your account. Please, try again");
                         continue;
                     }
                     try {
-                        BigDecimal newMoneyInAccount = moneyInAccount.subtract(BigDecimal.valueOf(Integer.parseInt(moneyToWithdraw)));
+                        BigDecimal newMoneyInAccount = moneyInAccount.subtract(BigDecimal.valueOf(Double.parseDouble(moneyToWithdraw)));
                         Writer output = new FileWriter(System.getProperty("user.dir") + "/" + Extranet.getCurrentUserName() + "/" + "acc" + userInput + ".txt", false);
                         output.write(String.valueOf(newMoneyInAccount) + "\r\n" + currencyInAccount.toUpperCase());
                         output.close();
                         if (Extranet.getMoneyATM() == null) {
-                            Extranet.setMoneyATM(BigDecimal.valueOf(Integer.parseInt(moneyToWithdraw)));
+                            Extranet.setMoneyATM(BigDecimal.valueOf(Double.parseDouble(moneyToWithdraw)));
                         } else {
-                            Extranet.setMoneyATM(Extranet.getMoneyATM().add(BigDecimal.valueOf(Integer.parseInt(moneyToWithdraw))));
+                            Extranet.setMoneyATM(Extranet.getMoneyATM().add(BigDecimal.valueOf(Double.parseDouble(moneyToWithdraw))));
                         }
                         Extranet.setCurrencyInATM(currencyInAccount.toUpperCase());
                         System.out.println("Money successfully withdrew");
@@ -228,7 +227,7 @@ public class Account {
         Scanner scan = new Scanner(System.in);
         String moneyToTransfer = scan.nextLine();
 
-        if (Integer.parseInt(moneyToTransfer) > Extranet.getMoneyATM().intValue())
+        if (Double.parseDouble(moneyToTransfer) > Extranet.getMoneyATM().doubleValue())
         {
             System.out.println("You want to transfer more money that you have in ATM");
         }
@@ -239,10 +238,10 @@ public class Account {
             String usernameToTransferMoney = scan.nextLine();
             if (newdb.checkUsername(usernameToTransferMoney)) {
                 List<String> lines = Files.readAllLines(Path.of(System.getProperty("user.dir") + "/" + usernameToTransferMoney + "/" + "acc1.txt"));
-                BigDecimal moneyInAccount = BigDecimal.valueOf(Integer.parseInt(lines.get(0)));
+                BigDecimal moneyInAccount = BigDecimal.valueOf(Double.parseDouble(lines.get(0)));
                 String currencyInAccount = lines.get(1);
                 if (currencyInAccount.equals(Extranet.getCurrencyInATM())) {
-                    BigDecimal newMoney = moneyInAccount.add(BigDecimal.valueOf(Integer.parseInt(moneyToTransfer)));
+                    BigDecimal newMoney = moneyInAccount.add(BigDecimal.valueOf(Double.parseDouble(moneyToTransfer)));
                     Writer output = new FileWriter(System.getProperty("user.dir") + "/" + usernameToTransferMoney + "/" + "acc1.txt", false);
                     output.write(String.valueOf(newMoney) + "\r\n" + currencyInAccount.toUpperCase());
                     output.close();
